requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
numpy>=1.26.0
av>=12.0.0
faster-whisper>=1.0.3
funasr>=1.1.14
openai-whisper>=20231117
PyYAML>=6.0.1
torch>=2.2.0
torchaudio>=2.2.0
transformers>=4.38.0
huggingface-hub>=0.23.0
modelscope>=1.18.0
mecab-python3>=1.0.8
unidic-lite>=1.0.8
jieba>=0.42.1
pypinyin>=0.50.0
num2words>=0.5.13
langid>=1.1.6
soundfile>=0.12.1
gruut[de,es,fr]>=2.2.3
